Refactor to use cogs
