--- frontroomsbot/__init__.py
@@ -1,6 +1,0 @@
-from collections import deque
-
-
-class A:
-    def __init__(self):
-        self.q = deque()
