+++ frontroomsbot/consts.py
@@ -7,4 +7,4 @@ GUILD = os.getenv("GUILD_ID")
 HF_TOKEN = os.getenv("HF_TOKEN")
 DB_CONN = os.getenv("DB_CONN")
 
-COGS_DIR = "frontroomsbot/cogs"
+COGS_DIR = "cogs"
